Issue #51: Delete util.py and move the b58encode function to meat.py which is the only place that uses it
